+++ src/console/InteractionController.ts
@@ -50,7 +50,7 @@ export class InteractionController {
 
     // Global options
     command = command
-      .concat(this.saveOptionIfExists(answers, '-S ', InteractiveQuestion.IGNORE_SOURCES))
+      .concat(this.saveOptionIfExists(answers, '-S ', InteractiveQuestion.SOURCES))
       .concat(this.saveOptionIfExists(answers, '-E ', InteractiveQuestion.IGNORE_EXTENSIONS))
       .concat(this.saveOptionIfExists(answers, '-m ', InteractiveQuestion.GRADUATE_OPERATIONS))
       .concat(this.saveOptionIfExists(answers, '-O ', InteractiveQuestion.LOG_FILENAME))